Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/script.https.sub-expected.txt

Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/echo-as-script.py
Blocked access to external URL https://www.127.0.0.1:9443/fetch/metadata/resources/echo-as-script.py
Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/echo-as-script.py
Blocked access to external URL https://www.127.0.0.1:9443/fetch/metadata/resources/echo-as-script.py

PASS Same-origin script
FAIL Same-origin script: sec-fetch-dest assert_equals: expected "script" but got ""
FAIL Same-origin script: sec-fetch-mode assert_equals: expected "no-cors" but got ""
FAIL Same-origin script: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS Same-origin script: sec-fetch-user
PASS Same-site script
FAIL Same-site script: sec-fetch-dest null is not an object (evaluating 'value.dest')
FAIL Same-site script: sec-fetch-mode null is not an object (evaluating 'value.mode')
FAIL Same-site script: sec-fetch-site null is not an object (evaluating 'value.site')
FAIL Same-site script: sec-fetch-user null is not an object (evaluating 'value.user')
PASS Cross-site script
FAIL Cross-site script: sec-fetch-dest null is not an object (evaluating 'value.dest')
FAIL Cross-site script: sec-fetch-mode null is not an object (evaluating 'value.mode')
FAIL Cross-site script: sec-fetch-site null is not an object (evaluating 'value.site')
FAIL Cross-site script: sec-fetch-user null is not an object (evaluating 'value.user')
PASS Same-origin CORS script
FAIL Same-origin CORS script: sec-fetch-dest assert_equals: expected "script" but got ""
FAIL Same-origin CORS script: sec-fetch-mode assert_equals: expected "cors" but got ""
FAIL Same-origin CORS script: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS Same-origin CORS script: sec-fetch-user
